Update watching.js
This is fix for an issue that was logged at: https://github.com/tailwindlabs/tailwindcss/issues/14632.

For folders that have the characters [, ], (, ), {, or }, the --watch option doesn't work. This is because for some reason, chokidar does not fire the .on('change') event when tailwind is watching from such directories. However, it fires the .on('raw') event with an evt type of 'change'. This works the same way for .on('add') events. Fix is to remove direct on('change') and on('add') event handling and handle both these events in the on('raw') event handling.

To reproduce the bug that this fixes, you can view my repo here that has instructions: https://github.com/lusayo-nyondo/tailwindcss_bug_report_sayo.